Update isOverlapTimeSlot to check entire timetable
