+++ src/main/java/seedu/zoomaster/command/planner/AddMeetingCommand.java
@@ -47,7 +47,7 @@ public class AddMeetingCommand extends AddSlotCommand {
     }
 
     @Override
-    protected String create(String command, Module module) throws ZoomasterException {
+    protected String create(String command, Module module, Timetable timetable) throws ZoomasterException {
         String message = "";
         List<String> slotAndBookmark = Arrays.asList(command.trim().split(" "));
         if (isAddModuleBookmark(slotAndBookmark)) {
@@ -65,10 +65,9 @@ public class AddMeetingCommand extends AddSlotCommand {
                         + slotAndBookmark.get(2) + " " + slotAndBookmark.get(3) + ") Please check format.");
             }
             Slot newSlot;
-            if (module.slotExists(lesson, day, startTime, endTime)) {
+            if (timetable.isOverlapTimeSlot(day, startTime, endTime)) {
                 newSlot = module.getSlot(lesson, day, startTime, endTime);
-                message +=  "  " + lesson + " slot already exists\n";
-                message += checkForAndAddBookmarkToSlot(slotAndBookmark, lesson, newSlot);
+                message +=  "  " + lesson + " overlaps with an existing timeslot\n";
             } else if (isAvailable(day, startTime, endTime)) {
                 newSlot = module.createSlotNew(lesson, day, startTime, endTime);
                 module.addSlot(newSlot);
